requirements.txt
fastapi
uvicorn
databases
asyncpg
pydantic
fastapi-analytics
python-dotenv
